changed development and testing environments to sqlite
I do not have postgres servers up and running so it's putting a hold on
development which I don't want. For the time being I changed the dialect
to sqlite.

I also added more logging just to make sure that the bot can interpret
messages sent. Did this because Idk whether the bot is just not doing
anything so having that there for the time being is quite useful to me.