fixed issue where belongsToMany would create keys by itself
`belongsToMany()` would create columns where it was just not needed like
UserUserId and GuildGuildId. By explicitly stating the foreign keys in
the function it seemed to have fixed the issue.

The database creation query also seems correct with no issues found
there.

The only conern I have with this change is how seequelize will handle
it. Will probably have to write some tests to make sure this is
consistent.