check if .env file exists
This was added to ensure that the env file exists before the bot is ran.

The reason why the process exits without throwing an Error is because it
is not an Error, it a misconfiguration. This is intended behaviour and
therefore I do not find it necessary to include callback and traces in
the logger. This isn't to say this might not change since there might be
fatal errors that need this information in order to be fixed.