created .env verifier
Moved the if statement that determined whether the environment variable
exists to this verifyDotEnv function.

This function was created to abstract the process of checking the
contents of the .env file. It was placed in the source as it own
distinct module because creating a submodule for it is overcomplicated.
Therefore I did not group it with any other files.