written tests for plus/minus two function
I'm committing this because of a weird bug where you can only plus/minus
two once before the database blows up. I'm quick to think that this
might be a problem with sequelize, but I also feel like the code isn't
that great. There really isn't any reason why this should be happening.

I will try to replicate this bug in a sandbox environment.