Refactored code to satisfy deno language server
Using the deno language server to code, so when I got the codebase there
were a bunch of errors. I didn't want to reconfigure the language server
so I changed some lines. Most notibly making all imports end with .js if
they are local and importing the process module instead of calling it
globally.