starting refactoring tests
For some reason the tests do not work on my local machine anymore. I
don't really know why this is the case, but I'm having alot of problem
when it comes to testing the database on all machines.

I'm going to try solve this by refactoring a little bit and seeing if
the behaviour changes.

It's hard writing tests when half of the time the tests don't work, are
machine dependent and do not give you a clue about what migt be going
on. :(